ToList 메서드 추가

+++ Kosdas/Models/PriceRecord.cs
@@ -3,7 +3,7 @@
     /// <summary>
     ///     가격 정보
     /// </summary>
-    public record PriceRecord(DateTime Date, decimal Open, decimal High, decimal Low, decimal Close, decimal Volume)
+    public record PriceRecord(DateTime Date, double Open, double High, double Low, double Close, double Volume)
     {
         public override string ToString() => $"{Date:d} {Open:N0} -> {Close:N0}";
     }
