+++ Kosdas/NaverPriceLoader.cs
@@ -36,11 +36,11 @@ namespace Kosdas
 
             return new PriceRecord(
                 DateTime.ParseExact(tokens[0], "yyyyMMdd", null),
-                Convert.ToDecimal(tokens[1]),
-                Convert.ToDecimal(tokens[2]),
-                Convert.ToDecimal(tokens[3]),
-                Convert.ToDecimal(tokens[4]),
-                Convert.ToDecimal(tokens[5])
+                Convert.ToDouble(tokens[1]),
+                Convert.ToDouble(tokens[2]),
+                Convert.ToDouble(tokens[3]),
+                Convert.ToDouble(tokens[4]),
+                Convert.ToDouble(tokens[5])
             );
         }
     }
